pyproject.toml
[tool.poetry]
name = "dvlife-collector"
version = "0.1.0"
description = ""
authors = ["tefuna <tefuna@gmail.com>"]
readme = "README.md"
packages = [{ include = "dvlife_collector" }]

[tool.poetry.dependencies]
python = "^3.11.2"
pyyaml = "^6.0"
numpy = "^1.25.1"
selenium = "^4.10.0"
beautifulsoup4 = "^4.12.2"
python-dotenv = "^1.0.0"
pandas = "^2.0.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
sphinx = "^7.0.1"
taskipy = "^1.11.0"

[tool.black]
line-length = 119

[tool.pytest.ini_options]
pythonpath = "dvlife_collector"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
position = "python dvlife_collector/main_position.py"
test = "poetry run pytest -v"
